Code cleanup
-Removed scanf/sscanf compiler warrnings
-Fixed some integer casting to avoid compiler warnings
-Do range check on starting private key value
-Fixed commandline processing in AddrGen.exe

+++ cudaInfo/main.cpp
@@ -10,16 +10,16 @@ void printDeviceInfo(const cuda::CudaDeviceInfo &info)
 	printf("Capability:  %d.%d\n", info.major, info.minor);
 	printf("MP:          %d\n", info.mpCount);
 	printf("Cores:       %d (%d per MP)\n", info.mpCount * info.cores, info.cores);
-	printf("Memory:      %dMB\n", info.mem / (1024 * 1024));
+	printf("Memory:      %dMB\n", (int)(info.mem / (1024 * 1024)));
 }
 
 int main(int argc, char **argv)
 {
 	std::vector<cuda::CudaDeviceInfo> devices = cuda::getDevices();
 
-	printf("Found %d devices\n\n", devices.size());
+	printf("Found %d devices\n\n", (int)devices.size());
 
-	for(int i = 0; i < devices.size(); i++) {
+	for(int i = 0; i < (int)devices.size(); i++) {
 		printDeviceInfo(devices[i]);
 		printf("\n");
 	}
